common report code for get or post handler to call
