+++ lib/handlers/index.js
@@ -1,0 +1,9 @@
+import getUserReportsHandler from './getUserReportsHandler';
+import postUserReportsHandler from './postUserReportsHandler';
+import userDataHandler from './userDataHandler';
+
+export default {
+  postUserReport: postUserReportsHandler,
+  getUserReport: getUserReportsHandler,
+  getUserData: userDataHandler,
+};
