+++ lib/handlers/postUserReportsHandler.js
@@ -1,0 +1,75 @@
+import fs from 'fs';
+import {
+  getSessionHeader,
+  createCounter,
+  exportTimeout,
+  logMaker,
+} from '../utils';
+
+import generateReport from '../reportUtils';
+
+const reportStatusCount = createCounter(
+  'tidepool_export_report_status_count',
+  'The number of errors for each status code.',
+  ['status_code', 'report_params'],
+);
+
+const log = logMaker('postUserReportsHandler.js', {
+  level: process.env.DEBUG_LEVEL || 'info',
+});
+
+export default function postUserReportsHandler() {
+  return async (req, res) => {
+    // Set the timeout for the request. Make it 10 seconds longer than
+    // our configured timeout to give the service time to cancel the API data
+    // request, and close the outgoing data stream cleanly.
+    req.setTimeout(exportTimeout + 10000);
+
+    try {
+      const userId = req.params.userid;
+      const {
+        dob,
+        fullName,
+        mrn,
+        bgUnits,
+        tzName,
+        restricted_token: restrictedToken,
+        reports,
+      } = req.query;
+
+      const timer = setTimeout(() => {
+        res.emit('timeout', exportTimeout);
+      }, exportTimeout);
+
+      const pdfReport = await generateReport(
+        log,
+        {
+          userId,
+          fullName,
+          dob,
+          mrn,
+        },
+        { tzName, bgUnits, reports },
+        { token: restrictedToken, sessionHeader: getSessionHeader(req) },
+      );
+
+      res.setHeader('Content-Disposition', 'attachment: filename="report.pdf"');
+      res.setHeader('Content-Type', 'application/octet-stream');
+      const blobArrayBuffer = await pdfReport.blob.arrayBuffer();
+      const pdfBuffer = Buffer.from(blobArrayBuffer);
+      fs.writeFileSync('test.pdf', pdfBuffer);
+      res.send(pdfBuffer);
+      clearTimeout(timer);
+    } catch (error) {
+      if (error.response && error.response.status === 403) {
+        reportStatusCount.inc({ status_code: 403, report_params: req.query });
+        res.status(403);
+        log.error(`403: ${error}`);
+      } else {
+        reportStatusCount.inc({ status_code: 500, report_params: req.query });
+        res.status(500);
+        log.error(`500: ${error}`);
+      }
+    }
+  };
+}
