+++ lib/handlers/userDataHandler.js
@@ -2,8 +2,8 @@ import axios from 'axios';
 import queryString from 'query-string';
 import dataTools from '@tidepool/data-tools';
 import {
-  getHeaders, createCounter, exportTimeout, logMaker, mmolLUnits,
-} from './utils';
+  getSessionHeader, createCounter, exportTimeout, logMaker, mmolLUnits,
+} from '../utils';
 
 const dataStatusCount = createCounter(
   'tidepool_export_status_count',
@@ -47,7 +47,7 @@ export default function userDataHandler() {
     try {
       const cancelRequest = axios.CancelToken.source();
 
-      const requestConfig = getHeaders(req);
+      const requestConfig = getSessionHeader(req);
       requestConfig.responseType = 'stream';
       requestConfig.cancelToken = cancelRequest.token;
       const dataResponse = await axios.get(
