+++ lib/reportUtils.js
@@ -1,11 +1,28 @@
 import moment from 'moment-timezone';
 import _ from 'lodash';
 import axios from 'axios';
-import fs from 'fs';
 import { generateAGPSVGDataURLS } from '@tidepool/viz/dist/genAGPURL';
-import { mgdLUnits, mmolLUnits } from './utils';
+import { DataUtil } from '@tidepool/viz/dist/data';
+import { Blob } from 'buffer';
+import PDFKit from './pdfkit';
+import {
+  fetchUserData, getServerTime, mgdLUnits, mmolLUnits,
+} from './utils';
 
-export const reportDataTypes = [
+const {
+  createPrintPDFPackage,
+  utils: PrintPDFUtils,
+} = require('@tidepool/viz/dist/print.js');
+
+// These need to be 'require'd in order to have the global Blob available when parsed
+const blobStream = require('blob-stream');
+
+global.Blob = Blob;
+
+PrintPDFUtils.PDFDocument = PDFKit;
+PrintPDFUtils.blobStream = blobStream;
+
+const reportDataTypes = [
   'cbg',
   'smbg',
   'basal',
@@ -225,7 +242,7 @@ function buildDataQueries(
  *    getPumpSettingsUploadRecordById: object|null
  * }}
  */
-export function getQueryOptions(data, units, serverTime, restrictedToken) {
+function getQueryOptions(data, units, serverTime, restrictedToken) {
   const options = { initial: true };
 
   if (restrictedToken) {
@@ -285,7 +302,7 @@ export function getQueryOptions(data, units, serverTime, restrictedToken) {
  * @param {array} data
  * @returns {{timezoneAware:boolean, timezoneName: string}} timePrefs
  */
-export function setTimePrefs(tzName, data) {
+function setTimePrefs(tzName, data) {
   const timePrefs = {
     timezoneAware: true,
     timezoneName: 'UTC',
@@ -312,12 +329,7 @@ async function graphRendererOrca(data, config) {
 /**
  *
  * @param {Array<object>} data user data
- * @returns {{
- *  agp: { startDate: string, endDate: string },
- *  basics: { startDate: string, endDate: string },
- *  bgLog: { startDate: string, endDate: string },
- *  daily: { startDate: string, endDate: string },
- *  }} startDate and endDate for the given report type
+ * @returns {ReportDates} startDate and endDate for the given report type
  */
 function getDateRangeByReport(
   data,
@@ -393,31 +405,14 @@ function getDateRangeByReport(
 
 /**
  *
- * @param {object} res http response
- * @param {object} pdf pdf report data to send
- */
-export async function sendPDFReport(res, pdf) {
-  res.setHeader('Content-Disposition', 'attachment: filename="report.pdf"');
-  res.setHeader('Content-Type', 'application/octet-stream');
-
-  const blobArrayBuffer = await pdf.blob.arrayBuffer();
-  const pdfBuffer = Buffer.from(blobArrayBuffer);
-  fs.writeFileSync('test.pdf', pdfBuffer);
-  res.send(pdfBuffer);
-}
-
-/**
- *
  * @param {*} queries
  * @param {*} options
  * @param {*} dataUtil
  * @returns {{ pdfData: object, options: object }}
  */
-export function getReportData(queries, opts, dataUtil) {
+function getReportData(queries, opts, dataUtil) {
   const containsDataForReport = (data, vals) => {
-    const dataVals = _.flatten(
-      _.valuesIn(_.get(data, vals, {})),
-    );
+    const dataVals = _.flatten(_.valuesIn(_.get(data, vals, {})));
     return dataVals.length > 0;
   };
   const containsDataForBasics = (data) => {
@@ -475,7 +470,7 @@ export function getReportData(queries, opts, dataUtil) {
   };
 }
 
-export function getReportOptions(
+function getReportOptions(
   data,
   bgUnits,
   reports = [allReports],
@@ -542,7 +537,7 @@ export function getReportOptions(
  * @param {*} agpPDFData
  * @returns {array} processedSVGs
  */
-export async function processAGPSVGs(agpPDFData) {
+async function processAGPSVGs(agpPDFData) {
   const agpSVGFigures = await generateAGPSVGDataURLS({
     ...agpPDFData,
   });
@@ -561,4 +556,118 @@ export async function processAGPSVGs(agpPDFData) {
   });
   const processedEntries = await Promise.all(agpFigurePromises);
   return _.fromPairs(processedEntries);
+}
+
+/**
+ * @typedef {{
+ *  agp: { startDate: string, endDate: string },
+ *   basics: { startDate: string, endDate: string },
+ *   bgLog: { startDate: string, endDate: string },
+ *   daily: { startDate: string, endDate: string },
+ * }} ReportDates
+ */
+
+/**
+ *
+ * @param {{
+ *   userId: string;
+ *   fullName: string;
+ *   dob: string;
+ *   mrn: string;
+ *  }} userDetail
+ * @param {{
+ *   tzName: string;
+ *   bgUnits: string;
+ *   reports: Array;
+ *   dates: ReportDates;
+ *  }} reportDetail
+ * @param {{
+ *   token: string;
+ *   sessionHeader: object;
+ *  }} requestData;
+ * @returns {object} pdfData
+ */
+export default async function generateReport(
+  log,
+  userDetail,
+  reportDetail,
+  requestData,
+) {
+  const dataUtil = new DataUtil();
+  const {
+    userId, fullName, dob, mrn,
+  } = userDetail;
+
+  const { tzName, bgUnits, reports } = reportDetail;
+
+  const { token, sessionHeader } = requestData;
+
+  const serverTime = await getServerTime();
+  log.debug('get server time ', serverTime);
+
+  const latestDatums = await fetchUserData(
+    userId,
+    {
+      type: reportDataTypes.join(','),
+      latest: 1,
+      endDate: moment.utc(serverTime).add(1, 'days').toISOString(),
+    },
+    sessionHeader,
+  );
+
+  const queryOptions = getQueryOptions(
+    latestDatums,
+    bgUnits,
+    serverTime,
+    token,
+  );
+  log.debug('data query options ', queryOptions);
+  const timePrefs = setTimePrefs(tzName, latestDatums);
+  log.debug('user timePrefs ', timePrefs);
+
+  const cancelRequest = axios.CancelToken.source();
+  const requestConfig = { headers: sessionHeader };
+  requestConfig.cancelToken = cancelRequest.token;
+  requestConfig.params = queryOptions;
+
+  const userData = await fetchUserData(
+    userId,
+    { ...queryOptions },
+    requestConfig.headers,
+  );
+  log.debug(`Downloading data for User ${userId}...`);
+
+  log.debug('add data to dataUtil');
+  const data = dataUtil.addData(userData, userId, false);
+  log.debug('getting report options');
+  const { queries, printOptions } = getReportOptions(
+    data,
+    bgUnits,
+    reports,
+    timePrefs,
+    userId,
+    {
+      fullName,
+      patient: {
+        mrn,
+        birthday: dob,
+      },
+    },
+  );
+
+  log.debug('getting pdf report data');
+  const reportData = getReportData(queries, printOptions, dataUtil);
+
+  if (!reportData.options.agp.disabled) {
+    reportData.options.svgDataURLS = await processAGPSVGs(
+      reportData.pdfData.agp,
+    );
+  }
+
+  const pdf = await createPrintPDFPackage(
+    reportData.pdfData,
+    reportData.options,
+  ).catch((error) => log.error(error));
+  log.debug('success', pdf);
+  return pdf;
 }
