+++ lib/utils.js
@@ -16,7 +16,7 @@ export const createCounter = (name, help, labelNames) => new Counter({
   registers: [register],
 });
 
-export function getHeaders(request) {
+export function getSessionHeader(request) {
   if (request.headers['x-tidepool-session-token']) {
     return {
       'x-tidepool-session-token': request.headers['x-tidepool-session-token'],
